feat: Add `Dictionary Corde` to get dictionary entity.
