+++ src/store/modules/ADempiere/formDefinition.js
@@ -25,7 +25,7 @@ import lang from '@/lang'
 import { CONTAINER_FORM_PREFIX } from '@/utils/ADempiere/dictionary/form/index.js'
 
 // API Request Methods
-import { requestForm } from '@/api/ADempiere/dictionary/index.ts'
+import { requestForm } from '@/api/ADempiere/dictionary/form'
 
 // Utils and Helper Methods
 import { showMessage } from '@/utils/ADempiere/notification'
@@ -89,7 +89,7 @@ const form = {
             const newForm = {
               ...formResponse,
               containerUuid: formResponse.uuid,
-              containerKey: CONTAINER_FORM_PREFIX + formResponse.id
+              containerKey: CONTAINER_FORM_PREFIX + formResponse.internal_id
             }
 
             commit('addForm', newForm)
